update to readFile for previous changes

+++ Assignment 2/src/mru/tsc/controller/GameManager.java
@@ -99,25 +99,25 @@ public class GameManager {
 						
 				if (firstDigit == ZERO || firstDigit == ONE) {
 					
-					Figure f = new Figure (Integer.parseInt(splittedLine[ZERO]), splittedLine[ONE], splittedLine[TWO], 
+					Figure f = new Figure (Long.parseLong(splittedLine[ZERO]), splittedLine[ONE], splittedLine[TWO], 
 							Double.parseDouble(splittedLine[THREE]), Integer.parseInt(splittedLine[FOUR]), 
 							splittedLine[FIVE], splittedLine[SIX].charAt(ZERO));
 					toys.add(f);
 				}else if (firstDigit == THREE || firstDigit == FOUR){
 					
-					Animal a = new Animal (Integer.parseInt(splittedLine[ZERO]), splittedLine[ONE], splittedLine[TWO], 
+					Animal a = new Animal (Long.parseLong(splittedLine[ZERO]), splittedLine[ONE], splittedLine[TWO], 
 							Double.parseDouble(splittedLine[THREE]), Integer.parseInt(splittedLine[FOUR]), 
 							splittedLine[FIVE], splittedLine[SIX].charAt(ZERO));
 					toys.add(a);
 				}else if (firstDigit == FOUR || firstDigit == FIVE || firstDigit == SIX){
 					
-					Puzzle p = new Puzzle (Integer.parseInt(splittedLine[ZERO]), splittedLine[ONE], splittedLine[TWO], 
+					Puzzle p = new Puzzle (Long.parseLong(splittedLine[ZERO]), splittedLine[ONE], splittedLine[TWO], 
 							Double.parseDouble(splittedLine[THREE]), Integer.parseInt(splittedLine[FOUR]), 
 							splittedLine[FIVE], splittedLine[SIX]);
 					toys.add(p);
 				}else {
 					
-					BoardGame b = new BoardGame (Integer.parseInt(splittedLine[ZERO]), splittedLine[ONE], splittedLine[TWO], 
+					BoardGame b = new BoardGame (Long.parseLong(splittedLine[ZERO]), splittedLine[ONE], splittedLine[TWO], 
 							Double.parseDouble(splittedLine[THREE]), Integer.parseInt(splittedLine[FOUR]), 
 							splittedLine[FIVE], splittedLine[SIX], splittedLine[SEVEN]);
 					toys.add(b);
